@ v1.7.186 — apply clippy machine-applicable idiom fixes (audit cleanup)
Whole-codebase integrity audit found no correctness bugs / no exploitable vulns. clippy reported 113 idiom/style warnings (0 bugs); applied the machine-applicable subset via cargo clippy --fix across ~28 Rust files (clamp, sort_by_key, .first(), is_some_and, matches!, flatten, split_once, drop redundant return/format/closures). Semantically equivalent, validated by cargo check + full test suites. Remaining ~58 are non-auto-fixable design notes (too-many-args, manual_clamp). The 2 cargo-audit advisories (sqlx 0.8.0, rsa 0.9.10) are transitive in the MySQL/Postgres path Lucy's SQLite usage never hits; the sqlx bump is blocked by a libsqlite3-sys link conflict with rusqlite 0.31.

+++ src-tauri/src/commands/script_verify.rs
@@ -317,7 +317,7 @@ pub async fn verify_script(language: String, content: String) -> Result<VerifyRe
             format!("spawn_blocking failed: {}", e), None, 0)),
         Err(_) => Ok(VerifyResult::err(&lang_label,
             format!("verify timed out after {}s", PROC_TIMEOUT_SECS + 1), None,
-            ((PROC_TIMEOUT_SECS + 1) * 1000) as u64)),
+            (PROC_TIMEOUT_SECS + 1) * 1000)),
     }
 }
